+++ Sources/SpeziOnboarding/ConsentView.swift
@@ -6,6 +6,7 @@
 // SPDX-License-Identifier: MIT
 //
 
+import CoreGraphics
 import PencilKit
 import SpeziViews
 import SwiftUI
@@ -42,20 +43,21 @@ public struct ConsentView<ContentView: View, Action: View>: View {
         }
     }
     
-    @EnvironmentObject var onboardingDataSource: OnboardingDataSource
+    
+    @EnvironmentObject private var onboardingDataSource: OnboardingDataSource
 
     private let contentView: ContentView
     private let action: Action
     private let givenNameField: FieldLocalizationResource
     private let familyNameField: FieldLocalizationResource
-    @State var name = PersonNameComponents()
+    private var exportConsentForm: Bool = false
+    private var asyncMarkdown: (() async -> Data)?
+    @State private var name = PersonNameComponents()
     @State private var showSignatureView = false
     @State private var isSigning = false
-    @State var signature = PKDrawing()
-    @State var signatureSize: CGSize = .zero
+    @State private var signature = PKDrawing()
+    @State private var signatureSize: CGSize = .zero
     
-    // TODO: Workaround to store the markdown here as otherwise no access to it for the export function
-    var asyncMarkdown: (() async -> Data)?
     
     public var body: some View {
         ScrollViewReader { proxy in
@@ -83,15 +85,6 @@ public struct ConsentView<ContentView: View, Action: View>: View {
                         
                         Divider()
                         
-                        // TODO: Just a placeholder to trigger the export functionality
-                        Button {
-                            Task {
-                                await export()
-                            }
-                        } label: {
-                            Text("Render")
-                        }
-                        
                         action
                             .disabled(buttonDisabled)
                             .animation(.easeInOut, value: buttonDisabled)
@@ -99,6 +92,14 @@ public struct ConsentView<ContentView: View, Action: View>: View {
                             .onChange(of: showSignatureView) { _ in
                                 proxy.scrollTo("ActionButtons")
                             }
+                            /// Use `.simultaneousGesture()` to detect tap as action subview would otherwise capture the tap gesture
+                            .simultaneousGesture(TapGesture().onEnded {
+                                if !buttonDisabled && exportConsentForm {
+                                    Task { @MainActor in
+                                        await self.export()
+                                    }
+                                }
+                            })
                     }
                     .transition(.opacity)
                     .animation(.easeInOut, value: showSignatureView)
@@ -121,13 +122,15 @@ public struct ConsentView<ContentView: View, Action: View>: View {
     
     
     /// Creates a ``ConsentView`` with a provided action view using  an``OnboardingActionsView`` and renders a markdown view.
+    /// Furthermore, by default, the signed consent form is exported to the `Standard` as a PDF.
     /// - Parameters:
     ///   - header: The header view will be displayed above the markdown content.
     ///   - asyncMarkdown: The markdown content provided as an UTF8 encoded `Data` instance that can be provided asynchronously.
     ///   - footer: The footer view will be displayed above the markdown content.
     ///   - action: The action that should be performed once the consent has been given.
-    ///   - givenNameField: The localization to use for the given (first) name field
-    ///   - familyNameField: The localization to use for the family (last) name field
+    ///   - givenNameField: The localization to use for the given (first) name field.
+    ///   - familyNameField: The localization to use for the family (last) name field.
+    ///   - exportConsentForm: Indicates weather the signed consent form should be exported as a PDF to the `Standard`. Defaults to true.
     public init(
         @ViewBuilder header: () -> some View = { EmptyView() },
         asyncMarkdown: @escaping () async -> Data,
@@ -135,49 +138,32 @@ public struct ConsentView<ContentView: View, Action: View>: View {
         action: @escaping () async -> Void,
         givenNameField: FieldLocalizationResource = LocalizationDefaults.givenName,
         familyNameField: FieldLocalizationResource = LocalizationDefaults.familyName,
-        export: Bool = true
+        exportConsentForm: Bool = true
     ) where ContentView == AnyView, Action == OnboardingActionsView {
-        // TODO
-        let contentView = AnyView(
-            VStack {
-                header()
-                DocumentView(
-                    asyncData: asyncMarkdown,
-                    type: .markdown
+        self.init(
+            contentView: {
+                AnyView(
+                    VStack {
+                        header()
+                        DocumentView(
+                            asyncData: asyncMarkdown,
+                            type: .markdown
+                        )
+                        footer()
+                    }
                 )
-                footer()
-            }
+            },
+            actionView: {
+                OnboardingActionsView(LocalizedStringResource("CONSENT_ACTION", bundle: .atURL(from: .module))) {
+                    await action()
+                }
+            },
+            givenNameField: givenNameField,
+            familyNameField: familyNameField
         )
         
-        if export {
-            self.init(
-                contentView: { contentView },
-                actionView: {
-                    OnboardingActionsView(
-                        primaryContent: .text(.init("CONSENT_ACTION", bundle: .atURL(from: .module))),
-                        primaryAction: { await action() },  // TODO
-                        secondaryContent: .image("square.and.arrow.up"),    // Share button for the export form
-                        secondaryAction: { await action() },    // TODO
-                        layout: .horizontal(proportions: 0.8)
-                    )
-                },
-                givenNameField: givenNameField,
-                familyNameField: familyNameField
-            )
-        } else {
-            self.init(
-                contentView: { contentView },
-                actionView: {
-                    OnboardingActionsView(LocalizedStringResource("CONSENT_ACTION", bundle: .atURL(from: .module))) {
-                        await action()
-                    }
-                },
-                givenNameField: givenNameField,
-                familyNameField: familyNameField
-            )
-        }
-        
         self.asyncMarkdown = asyncMarkdown
+        self.exportConsentForm = exportConsentForm
     }
 
     /// Creates a ``ConsentView`` with a provided action view using  an``OnboardingActionsView`` and renders HTML in a web view.
@@ -223,8 +209,8 @@ public struct ConsentView<ContentView: View, Action: View>: View {
     /// - Parameters:
     ///   - contentView: The content view providing context about the consent view.
     ///   - actionView: The action view that should be displayed under the name and signature boxes.
-    ///   - givenNameField: The localization to use for the given (first) name field
-    ///   - familyNameField: The localization to use for the family (last) name field
+    ///   - givenNameField: The localization to use for the given (first) name field.
+    ///   - familyNameField: The localization to use for the family (last) name field.
     public init(
         @ViewBuilder contentView: () -> ContentView,
         @ViewBuilder actionView: () -> Action,
@@ -237,6 +223,119 @@ public struct ConsentView<ContentView: View, Action: View>: View {
         self.familyNameField = familyNameField
     }
 }
+
+
+/// Extension of ``ConsentView`` enabling the export of the signed consent page in the onboarding flow.
+private extension ConsentView {
+    /// Creates a view representation of the consent content, ready for PDF export via SwiftUIs `ImageRenderer`.
+    /// At the moment, this is
+    ///
+    /// This function constructs a view for presenting the markdown consent form. It combines the
+    /// given markdown and the user's signature with details such as the date of export. It can be
+    /// used to create exportable PDF documents of the consent form.
+    ///
+    /// - Parameters:
+    ///   - markdown: The markdown consent content as an `AttributedString`.
+    ///
+    /// - Returns: A SwiftUI `View` representation of the consent content and signature.
+    ///
+    /// - Note: This function avoids the use of asynchronous operations.
+    /// Asynchronous tasks are incompatible with SwiftUI's `ImageRenderer`,
+    /// which expects all rendering processes to be synchronous.
+    func exportBody(markdown: AttributedString) -> some View {
+        VStack {
+            HStack {
+                Spacer()
+                
+                Text("Exported: \(DateFormatter.localizedString(from: Date(), dateStyle: .medium, timeStyle: .short))")
+                    .font(.caption)
+            }
+            .padding()
+            
+            Text("Spezi Consent")
+                .font(.title)
+            
+            Text(markdown)
+                .padding()
+            
+            Spacer()
+            
+            ZStack(alignment: .bottomLeading) {
+                Rectangle()
+                    .fill(.secondary)
+                    .frame(maxWidth: .infinity, maxHeight: 1)
+                    .padding(.horizontal, 20)
+                    .padding(.bottom, 30)
+                Text("X")
+                    .font(.title2)
+                    .foregroundColor(.secondary)
+                    .padding(.horizontal, 20)
+                    .padding(.bottom, 30 + 2)
+                Text(name.formatted(.name(style: .long)))
+                    .font(.subheadline)
+                    .foregroundColor(.secondary)
+                    .padding(.horizontal, 20)
+                    .padding(.bottom, 30 - 18)
+                Image(uiImage: signature.image(from: .init(x: 0, y: 0, width: signatureSize.width, height: signatureSize.height), scale: UIScreen.main.scale))
+            }
+            .frame(width: signatureSize.width, height: signatureSize.height)
+        }
+    }
+    
+    
+    /// Exports the consent form as a PDF in the specified paper size.
+    ///
+    /// This function retrieves the markdown content, renders it to an image, and saves it as a PDF
+    /// with the provided paper size. The resulting PDF is stored via the Spezi `Standard`.
+    /// The `Standard` must conform to the ``OnboardingConstraint``.
+    ///
+    /// - Parameter paperSize: The desired size for the exported PDF, defaulting to `.usLetter`.
+    func export(paperSize: PaperSize = .usLetter) async {
+        guard let asyncMarkdown else {
+            preconditionFailure("SpeziOnboarding: Consent form export is only supported for Markdown documents!")
+        }
+        
+        let markdown = await asyncMarkdown()
+        
+        let markdownString = (try? AttributedString(
+            markdown: markdown,
+            options: .init(interpretedSyntax: .inlineOnlyPreservingWhitespace)
+        )) ?? AttributedString(String(localized: "MARKDOWN_LOADING_ERROR", bundle: .module))
+        
+        let renderer = ImageRenderer(content: exportBody(markdown: markdownString))
+        let paperSize = CGSize(
+            width: paperSize.dimensions.width,
+            height: paperSize.dimensions.height
+        )
+        renderer.proposedSize = .init(paperSize)
+        
+        renderer.render { size, context in
+            var box = CGRect(origin: .zero, size: paperSize)
+            
+            /// Creates the `CGContext` that stores the to-be-rendered PDF in-memory as a Swift `Data` struct.
+            guard let mutableData = CFDataCreateMutable(kCFAllocatorDefault, 0),
+                  let consumer = CGDataConsumer(data: mutableData),
+                  let pdf = CGContext(consumer: consumer, mediaBox: &box, nil) else {
+                return
+            }
+            
+            pdf.beginPDFPage(nil)
+            pdf.translateBy(
+                x: 0,
+                y: 0
+            )
+            
+            context(pdf)
+            
+            pdf.endPDFPage()
+            pdf.closePDF()
+            
+            /// Stores the finished PDF within the Spezi `Standard`.
+            onboardingDataSource.store(mutableData as Data)
+        }
+    }
+}
+
 
 #if DEBUG
 struct ConsentView_Previews: PreviewProvider {
