+++ Sources/SpeziOnboarding/OnboardingConstraint.swift
@@ -15,4 +15,6 @@ public protocol OnboardingConstraint: Standard {
     /// Adds a new exported consent form represented as `Data` to the `Standard` conforming to ``OnboardingConstraint``.
     /// - Parameter consent: The exported consent form represented as `Data` that should be added.
     func store(consent: Data)
+    
+    func loadConsent() throws -> Data
 }
